uniprot-mapping-multiple-results-issue-solutionv1
When querying uniprot for a gene name, often the results include proteins that are encoded by different genes than the gene which was used for querying (eg. if I query "FLT1", a gene can be returned with the 'gene' property "RACK1" and thus not the gene I queried in the first place).

This update removes the option to cycle through such genes and only gives you the option to trust/cycle gene products which contain the correct gene.